requirements.txt
django>=5.0.0,<5.1.0
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers